Semantic Versioning + Ignore chunky (for now)
I don't know how date-based versioning resolves versioning as well as semantic, so sticking to semantic versioning.

Chunky won't see an initial release for a while; because it's a FS in the sense that it's a folder/file hierarchy, but unlike a FS, we're more interested in the FourCC's used to describe the folder/file than the name of the file. But I don't believe chunky requires these names to be unique, and the FourCCs are definitely not unique.